IDE updates, fix mouse hook hotkey
Added support for headerless (RDSK) HDD images, and fixed the IDE controller emulation to not keep separate LBA1-4 values for each hard drive.
This may look strange at first glance, but the Amiga would regularly write the CHS values for drive access ahead of actually switching to the drive it wanted, so I assume this is how it's actually supposed to work.
The second drive still doesn't mount automatically on boot, but I believe that some currently unsupported register write or read is causing this.