Initial alpha of the PiStorm GUI tool
It can only change the config file and enable/disable RTG right now, but
it a framework for adding more things in the future.

Some of the CLI tool code was slightly modified for VBCC compatibility,
most importantly, some macros that didn't work were expanded and some
includes were changed.