[MEGA-WIP] Raylib-based RTG output
NOTE: A working keyboard must be attached to the Raspberry Pi while testing this, otherwise it's impossible to actually quit the emulator.
raylib takes possession of the SSH keyboard for some reason, which makes it so you can't Ctrl+C out of the emulator over SSH, you must Ctrl+C or press Q on the Pi keyboard.

A mostly working RTG implementation using raylib instead of SDL2.0
Greatly decreases the rendering overhead for 8bpp modes and gets rid of the need for hardware ARGB888 texture format support.
RTG will be initialized using the resolution of the Raspberry Pi, and onbly the 320x200/320x240 modes are currently scaled to the full vertical area of the screen.