PiSCSI: Allow multiple partitions with the same name
Because people just can't stop naming the partitions on every single one of their hard drive images `DH0` and `DH1`.
If something doesn't work with this approach, I have no idea why, because as far as I can tell in the example boot ROM I have available to me, this is all that one does.